refactor(tests): utils changed to fake_devices
